refactor: rename folder from clients to integrations
